Add support for wise guys and pedantic assholes

+++ bot.py
@@ -72,6 +72,8 @@ bot = commands.Bot(command_prefix='!', description=description, intents=intents)
 # Implement common bot events
 #
 
+PATRONIZE_CMD = '!patronize'
+
 @bot.event
 async def on_ready():
     print('Logged in as', bot.user.name)
@@ -94,45 +96,47 @@ async def on_member_join(member):
 @bot.event
 async def on_message(message):
     print('message:', message.type, message.content)
-    # if bot is being mentioned
-    if bot.user.mentioned_in(message):
-        raw_text_i = message.clean_content.index(bot.user.name) + len(bot.user.name) + 1
-        raw_text = message.clean_content[raw_text_i:].strip()
-        cmd = raw_text.split()[0]
-        # basic health check
-        if cmd == 'ping':
-            await message.channel.send(message.author.mention + ' pong')
-        # print words list in channel
-        elif cmd == 'words':
-            # found this gem when looking up embeds:
-            # https://stackoverflow.com/questions/62390411/discord-py-discord-embed-from-dict-not-creating-an-embed-properly
-            words_embed = discord.Embed.from_dict(thesaurus.to_embed_dict())
-            await message.channel.send(embed=words_embed)
-        # print phrases list in channel
-        elif cmd == 'phrases':
-            phrases_embed = discord.Embed.from_dict(phrasebook.to_embed_dict())
-            await message.channel.send(embed=phrases_embed)
-        # add new word to thesaurus
-        elif cmd == 'word':
-            new_word_i = raw_text.index(cmd) + len(cmd) + 1
-            new_word = raw_text[new_word_i:].strip()
-            thesaurus.add(new_word)
-            success_embed = discord.Embed.from_dict({
-                'description': f'Patronizor has added **{new_word}** to its thesaurus'
-            })
-            await message.channel.send(embed=success_embed)
-        # add new phrase to phrasebook
-        elif cmd == 'phrase':
-            new_phrase_i = raw_text.index(cmd) + len(cmd) + 1
-            new_phrase = raw_text[new_phrase_i:].strip()
-            phrasebook.add(new_phrase)
-            success_embed = discord.Embed.from_dict({
-                'description': f'Patronizor has added **{new_phrase}** to its phrasebook'
-            })
-            await message.channel.send(embed=success_embed)
-    # remove default Discord new member message
-    if message.type == discord.MessageType.new_member:
-        await message.channel.delete_messages([message])
+    # do not process !patronize command here
+    if PATRONIZE_CMD not in message.content:
+        # if bot is being mentioned
+        if bot.user.mentioned_in(message):
+            raw_text_i = message.clean_content.index(bot.user.name) + len(bot.user.name) + 1
+            raw_text = message.clean_content[raw_text_i:].strip()
+            cmd = raw_text.split()[0]
+            # basic health check
+            if cmd == 'ping':
+                await message.channel.send(message.author.mention + ' pong')
+            # print words list in channel
+            elif cmd == 'words':
+                # found this gem when looking up embeds:
+                # https://stackoverflow.com/questions/62390411/discord-py-discord-embed-from-dict-not-creating-an-embed-properly
+                words_embed = discord.Embed.from_dict(thesaurus.to_embed_dict())
+                await message.channel.send(embed=words_embed)
+            # print phrases list in channel
+            elif cmd == 'phrases':
+                phrases_embed = discord.Embed.from_dict(phrasebook.to_embed_dict())
+                await message.channel.send(embed=phrases_embed)
+            # add new word to thesaurus
+            elif cmd == 'word':
+                new_word_i = raw_text.index(cmd) + len(cmd) + 1
+                new_word = raw_text[new_word_i:].strip()
+                thesaurus.add(new_word)
+                success_embed = discord.Embed.from_dict({
+                    'description': f'Patronizor has added **{new_word}** to its thesaurus'
+                })
+                await message.channel.send(embed=success_embed)
+            # add new phrase to phrasebook
+            elif cmd == 'phrase':
+                new_phrase_i = raw_text.index(cmd) + len(cmd) + 1
+                new_phrase = raw_text[new_phrase_i:].strip()
+                phrasebook.add(new_phrase)
+                success_embed = discord.Embed.from_dict({
+                    'description': f'Patronizor has added **{new_phrase}** to its phrasebook'
+                })
+                await message.channel.send(embed=success_embed)
+        # remove default Discord new member message
+        if message.type == discord.MessageType.new_member:
+            await message.channel.delete_messages([message])
     # forward command to normal processing
     await bot.process_commands(message)
 
@@ -149,6 +153,10 @@ async def patronize(ctx, username: str):
     # first check for username, then for mention
     member_by_name = discord.utils.find(lambda m: m.name.lower() == username.lower(), ctx.guild.members)
     member_to_patronize = member_by_name or (mentions[0] if 0 < len(mentions) else None) or None
+    if member_to_patronize == bot.user or username == bot.user.name:
+        # patronize author for even trying to outsmart me
+        await ctx.send(ctx.author.mention + ' nice try, ' + patronizing_word)
+        return
     if member_to_patronize is None:
         # patronize author if he fails to do this correctly
         await ctx.send(ctx.author.mention + ' you gotta patronize someone, ' + patronizing_word)
